runner/requirements_news.txt
Django==1.6.11
MySQL-python==1.2.5
Pillow==2.8.2
argparse==1.2.1
django-appconf==1.0.2
django-autocomplete-light==3.1.6
django-bootstrap-form==3.2
django-bootstrap3==5.4.0
django-forms-bootstrap==3.0.1
django-tagging==0.4.3
django-user-accounts==1.0.1
djangorestframework==3.1.1
imagestore==2.9.1
pinax-theme-bootstrap==5.7.5
psycopg2==2.6.1
pytz==2016.4
six==1.9.0
sorl-thumbnail==12.3
swapper==0.3.0
wsgiref==0.1.2
